[add] Hoang 12-9 - 2023 (3:27pm)

+++ Assets/_Data/Scripts/_ScripableObject/ItemDataSO.cs
@@ -6,7 +6,7 @@ public class ItemDataSO : ScriptableObject
 {
     [Header("COMMON")]
     public string ItemName;
-    public float ItemValue;
+    public int ItemValue;
     public ItemType ItemType;
     public GameObject Icon; //UI
     public GameObject Model;
